+++ src/pages/careers/_model/_locations/_locations.tsx
@@ -23,23 +23,23 @@ import {
     address_berlin,
 } from './_contact-details'
 import { LocationsType } from './_locations.types'
-import MalaysiaFlagIcon from 'images/svg/careers/career_flag_malaysia.svg'
-import ParaguayFlagIcon from 'images/svg/careers/career_flag_paraguay.svg'
-import UAEFlagIcon from 'images/svg/careers/career_flag_uae.svg'
-import MaltaFlagIcon from 'images/svg/careers/career_flag_malta.svg'
-import CyprusFlagIcon from 'images/svg/careers/career_flag_cyprus.svg'
-import CaymanFlagIcon from 'images/svg/careers/career_flag_cayman.svg'
-import RwandaFlagIcon from 'images/svg/careers/career_flag_rwanda.svg'
-import BVIFlagIcon from 'images/svg/careers/career_flag_bvi.svg'
-import UKFlagIcon from 'images/svg/careers/career_flag_uk.svg'
-import FranceFlagIcon from 'images/svg/careers/career_flag_france.svg'
-import ChannelIslandsFlagIcon from 'images/svg/careers/career_flag_channel_islands.svg'
-import GermanyFlagIcon from 'images/svg/careers/career_flag_germany.svg'
-import VanuatuFlagIcon from 'images/svg/careers/career_flag_vanuatu.svg'
-import JordanFlagIcon from 'images/svg/careers/career_flag_jordan.svg'
-import SingaporeFlagIcon from 'images/svg/careers/career_flag_singapore.svg'
+import MalaysiaFlagIcon from 'images/svg/flags/flag_malaysia.svg'
+import ParaguayFlagIcon from 'images/svg/flags/flag_paraguay.svg'
+import UAEFlagIcon from 'images/svg/flags/flag_uae.svg'
+import MaltaFlagIcon from 'images/svg/flags/flag_malta.svg'
+import CyprusFlagIcon from 'images/svg/flags/flag_cyprus.svg'
+import CaymanFlagIcon from 'images/svg/flags/flag_cayman.svg'
+import RwandaFlagIcon from 'images/svg/flags/flag_rwanda.svg'
+import BVIFlagIcon from 'images/svg/flags/flag_bvi.svg'
+import UKFlagIcon from 'images/svg/flags/flag_uk.svg'
+import FranceFlagIcon from 'images/svg/flags/flag_france.svg'
+import ChannelIslandsFlagIcon from 'images/svg/flags/flag_channel_islands.svg'
+import GermanyFlagIcon from 'images/svg/flags/flag_germany.svg'
+import VanuatuFlagIcon from 'images/svg/flags/flag_vanuatu.svg'
+import JordanFlagIcon from 'images/svg/flags/flag_jordan.svg'
+import SingaporeFlagIcon from 'images/svg/flags/flag_singapore.svg'
 import { localize } from 'components/localization'
-import HongKong from 'images/svg/careers/career_flag_hongkong.svg'
+import HongKong from 'images/svg/flags/flag_hongkong.svg'
 
 export const cyberjaya: LocationsType = {
     name: 'cyberjaya',
